Use the proper protoc and Arcus for building CuraEngine.
If you have protoc installed on your system, it will usually be used instead
of the one built by the build script, which can cause issues if it's
protoc v2.x for example. Also, if you have Arcus installed locally,
then it will force the use of the protobuf include/library that libArcus
was built with, which can be different from the one built by this script.
This causes errors in compilation because the cura-build protoc will use an API
that may not be available on your system protobuf.